+++ graphistry/plugins/__init__.py
@@ -1,3 +1,4 @@
 from .igraph import from_igraph, to_igraph
+from .gexf import from_gexf
 from .graphviz import layout_graphviz, render_graphviz
 from .cugraph import from_cugraph, to_cugraph, compute_cugraph, layout_cugraph